requirements.txt
# Core dependencies
pyyaml>=6.0
python-frontmatter>=1.0.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
types-PyYAML>=6.0.0  # Type stubs for mypy

# Documentation
# sphinx>=6.0.0  # Optional: for generating docs
